add commands switch statement
